Code Refactored - remove globalStyles directory
